14.01.21

+++ stm32h7/common_h7/Inc/system/hw_util.h
@@ -38,14 +38,14 @@ bool          HW_GetHWClockStatus     ( void *hw );
 void          HW_SetHWClock           ( void *hw, bool bOn );
 void          HW_SetDmaChClock        ( const HW_DmaType *tx, const HW_DmaType *rx);
 void          HW_Reset                ( void *hw );
-uint16_t      HW_GetLn2               ( uint16_t pwrof2 );
+uint16_t      HW_GetLn2               ( uint32_t pwrof2 );
 uint32_t *    HW_GetPeriphBitBandAddr ( __IO uint32_t *periphAddr, uint16_t bit_number );
 uint32_t      HW_GetGPIOIdx           (GPIO_TypeDef *gp);
 char          HW_GetGPIOLetter        (GPIO_TypeDef *gp);
 
 
 #define       HW_GetGPIO(a)           (GPIO_TypeDef *) HW_GetHW( COMBINE('G', a) );
-#define       HW_GetIdxFromPin(pin)   HW_GetLn2(pin)
+#define       HW_GetIdxFromPin(pin)   HW_GetLn2((uint16_t)pin)
 
 void          HW_DMA_HandleInit      (DMA_HandleTypeDef *hdma, const HW_DmaType *dma, void *parent );
